chore(cli): Rename to cedarPluralize
