+++ packages/cli/src/commands/generate/sdl/sdlHandler.js
@@ -8,13 +8,13 @@ import { generate as generateTypes } from '@cedarjs/internal/dist/generate/gener
 import { getConfig } from '@cedarjs/project-config'
 import { errorTelemetry } from '@cedarjs/telemetry'
 
+import { pluralize } from '../../../lib/cedarPluralize.js'
 import c from '../../../lib/colors.js'
 import { transformTSToJS, writeFilesTask } from '../../../lib/index.js'
 import {
   prepareForRollback,
   addFunctionToRollback,
 } from '../../../lib/rollback.js'
-import { pluralize } from '../../../lib/rwPluralize.js'
 import {
   getSchema,
   getEnum,
