+++ packages/cli/src/lib/cedarPluralize.js
