+++ packages/cli/src/lib/schemaHelpers.js
@@ -2,8 +2,8 @@ import prismaInternals from '@prisma/internals'
 
 import { getSchemaPath } from '@cedarjs/project-config'
 
+import { singularize, isPlural } from './cedarPluralize.js'
 import { ensureUniquePlural } from './pluralHelpers.js'
-import { singularize, isPlural } from './rwPluralize.js'
 
 import { getPaths } from './index.js'
